Update Kibana to version 4.5.X
In version 4+ of Kibana, Kibana now runs as its own service. Apache is
now configured to proxypass to the kibana application. This means that
the end user's browser does not need to connect directly to
ElasticSearch any longer. This change also moves to installing Kibana
from the elastic.co package repository instead of from a tarball.

This version of Kibana is dependent on ElasticSearch version >= 2.3.
Support Matrix URL
https://www.elastic.co/support/matrix#show_compatibility

Resolves: #1073